+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.1
+
+`showAndUploadToGitHub` now returns a `Future<Issue>`
+
+uploadToGitHub's screenshot is now optional
+
 ## 1.0.0
 
 initial release.